coreplan_visual_dashboard/pythonProject/requirements.txt
dash
dash-bootstrap-components
pandas
plotly
mysql-connector-python
python-dotenv
sqlalchemy
pymysql
cryptography